requirements.txt
plotly
nicegui
numpy
pyserial
opencv-python
Pillow
pytest
pytest-asyncio
watchdog